some fixs to the code and playing with the path
