+++ MotionProfiling_Project/src/main/java/frc/robot/PathCreater.java
@@ -1,9 +1,11 @@
 package frc.robot;
 
 import java.io.File;
-import frc.robot.RobotConstants.Path;
+import java.io.IOException;
+
 import jaci.pathfinder.Pathfinder;
 import jaci.pathfinder.Trajectory;
+import jaci.pathfinder.Waypoint;
 
 /**
  * This class generates all the paths for the path follower command to use.
@@ -28,6 +30,32 @@ public class PathCreater {
         for (Path path : Path.values()) {
             String s = "/home/lvuser/Paths " + path.name() + ".csv";
             Pathfinder.writeToCSV(new File(s), path.getTrajectory());
+        }
+    }
+    public enum Path {
+        SCALE(new Waypoint[] { new Waypoint(0, 0, 0), new Waypoint(2, -3, 0) });
+
+        private final Trajectory trajectory;
+
+        private Path(Waypoint[] path) {
+            trajectory = Pathfinder.generate(path, new Trajectory.Config(Trajectory.FitMethod.HERMITE_QUINTIC, Trajectory.Config.SAMPLES_HIGH,
+            RobotConstants.TIMEFRAME, RobotConstants.MAX_VELOCITY, RobotConstants.MAX_ACCELERATION,
+            RobotConstants.MAX_JERK));
+        }
+
+        private Path(File csvFile) {
+            Trajectory trajectory = null;
+            try {
+                trajectory = Pathfinder.readFromCSV(csvFile);
+            } catch (IOException e) {
+                System.err.println("File not existing");
+            }
+            this.trajectory = trajectory;
+
+        }
+
+        public Trajectory getTrajectory() {
+            return trajectory;
         }
     }
