+++ MotionProfiling_Project/src/main/java/frc/robot/RobotConstants.java
@@ -19,38 +19,14 @@ public class RobotConstants {
     public static final double MAX_ACCELERATION = 2;
     public static final double MAX_JERK = 80;
 
-    public static final PidSettings MOTION_PROFILING_PID_SETTINGS_LEFT = new PidSettings(0.3, 0, 0, 0);
-    public static final PidSettings MOTION_PROFILING_PID_SETTINGS_RIGHT = new PidSettings(0.3, 0, 0, 0);
+    public static final PidSettings MOTION_PROFILING_PID_SETTINGS_LEFT = new PidSettings(0.0, 0, 0.0, 0.0);//0.2103, 0.5139
+    public static final PidSettings MOTION_PROFILING_PID_SETTINGS_RIGHT = new PidSettings(0.0, 0, 0.0, 0.0);//0.1929, 2.5147
 
-    public static final double MOTION_PROFILING_KP_TURN = 0.8;
+    public static final double MOTION_PROFILING_KP_TURN = 0.0025;
 
     public static final double TICKS_PER_METER_RIGHT = 575.0;
     public static final double TICKS_PER_METER_LEFT = 771.5;
-    public static final int TICKS_PER_REVOLUTION_RIGHT = 50;
-    public static final int TICKS_PER_REVOLUTION_LEFT = 3;
-
-    public enum Path {
-        SCALE(new Waypoint[] { new Waypoint(0, 0, 0), new Waypoint(2, -3, 0) });
-
-        private final Trajectory trajectory;
-
-        private Path(Waypoint[] path) {
-            trajectory = Pathfinder.generate(path, Robot.pathCreater.config);
-        }
-
-        private Path(File csvFile) {
-            Trajectory trajectory = null;
-            try {
-                trajectory = Pathfinder.readFromCSV(csvFile);
-            } catch (IOException e) {
-                System.err.println("File not existing");
-            }
-            this.trajectory = trajectory;
-
-        }
-
-        public Trajectory getTrajectory() {
-            return trajectory;
-        }
-    }
+    public static final int TICKS_PER_REVOLUTION_RIGHT = 290;
+    public static final int TICKS_PER_REVOLUTION_LEFT = 360;
+    
 }
