+++ MotionProfiling_Project/src/main/java/frc/robot/SplitTrajectories.java
@@ -1,8 +1,8 @@
 package frc.robot;
 
+import frc.robot.PathCreater.Path;
 import jaci.pathfinder.Trajectory;
 import jaci.pathfinder.modifiers.TankModifier;
-import frc.robot.RobotConstants.Path;
 
 /**
  * This class splits the trajectories for each side of the robot.
@@ -13,6 +13,7 @@ public class SplitTrajectories {
     private Path path;
 
     public SplitTrajectories(Path path) {
+        this.path = path;
         this.modifier = new TankModifier(this.path.getTrajectory());
         this.modifier.modify(RobotConstants.WHEEL_BASE_WIDTH);
     }
